pyproject.toml
[project]
name = "fantasy_predictor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "football>=0.2.0",
    "fpl>=0.6.35",
    "langchain-openai>=0.3.27",
    "langgraph>=0.5.1",
    "pydantic>=2.11.7",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]
